@@ -1,20 +1,12 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="TagHelpers\**" />
-    <Content Remove="TagHelpers\**" />
-    <EmbeddedResource Remove="TagHelpers\**" />
-    <None Remove="TagHelpers\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Csla.AspNetCore" Version="5.1.0-R19110101" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="3.0.0" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.0.0" />
+    <PackageReference Include="Csla.AspNetCore" Version="6.0.0-R21120603" />
   </ItemGroup>
 
   <ItemGroup>
