@@ -4,8 +4,7 @@
     ViewData["Title"] = "Home page";
 }
 
-  <div class="text-center">
+<div class="text-center">
     <h1 class="display-4">CSLA .NET Razor Pages Example</h1>
-    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
     <p>Learn about <a href="https://cslanet.com">building apps with CSLA .NET</a>.</p>
-  </div>
+</div>
