@@ -1,24 +1,30 @@
-﻿using System.Threading.Tasks;
-using BusinessLibrary;
+﻿using BusinessLibrary;
 using Csla;
 using Csla.AspNetCore.RazorPages;
-using Microsoft.AspNetCore.Mvc;
 
 namespace RazorPagesExample.Pages.People
 {
   public class EditModel : PageModel<PersonEdit>
   {
+    public EditModel(ApplicationContext applicationContext, IDataPortal<PersonEdit> portal)
+      : base(applicationContext)
+    {
+      _portal = portal;
+    }
+
+    private IDataPortal<PersonEdit> _portal;
+
     public async Task OnGet(int id)
     {
       if (id == -1)
-        Item = await DataPortal.CreateAsync<PersonEdit>();
+        Item = await _portal.CreateAsync();
       else
-        Item = await DataPortal.FetchAsync<PersonEdit>(id);
+        Item = await _portal.FetchAsync(id);
     }
 
-    public async Task<ActionResult> OnPost()
+    public async Task OnPost()
     {
-      return await SaveAsync((Item.Id > -1));
+      await SaveAsync((Item.Id > -1));
     }
   }
 }
