@@ -1,22 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using BusinessLibrary;
+﻿using BusinessLibrary;
 using Csla;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace RazorPagesExample.Pages
 {
   public class ListPeopleModel : PageModel
   {
+    public ListPeopleModel(IDataPortal<PersonList> portal)
+    {
+      _portal = portal;
+    }
+
+    private readonly IDataPortal<PersonList> _portal;
 
     public PersonList PersonList { get; set; }
 
     public async Task OnGet()
     {
-      PersonList = await DataPortal.FetchAsync<PersonList>();
+      PersonList = await _portal.FetchAsync();
     }
   }
 }
