@@ -1,4 +1,4 @@
-@using RazorPagesExample
+﻿@using RazorPagesExample
 @namespace RazorPagesExample.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Csla.AspNetCore
