@@ -5,7 +5,6 @@
 // </copyright>
 // <summary>Application context manager that uses HttpContext</summary>
 //-----------------------------------------------------------------------
-using System;
 using Csla.Core;
 using System.Web;
 
@@ -17,24 +16,11 @@ namespace Csla.Web
   /// </summary>
   public class ApplicationContextManager : IContextManager
   {
-    private IServiceProvider _serviceProvider;
-    private ApplicationContext ApplicationContext { 
-      get 
-      {
-        return (ApplicationContext)_serviceProvider.GetService(typeof(ApplicationContext));
-      }
-    }
-
     /// <summary>
     /// Creates an instance of the type.
     /// </summary>
-    /// <param name="serviceProvider">Service provider instance.</param>
-    public ApplicationContextManager(IServiceProvider serviceProvider)
-    {
-      if (serviceProvider == null)
-        throw new ArgumentNullException(nameof(serviceProvider));
-      _serviceProvider = serviceProvider;
-    }
+    public ApplicationContextManager()
+    { }
 
     private const string _localContextName = "Csla.LocalContext";
     private const string _clientContextName = "Csla.ClientContext";
@@ -106,6 +92,23 @@ namespace Csla.Web
     public void SetClientContext(ContextDictionary clientContext, ApplicationContext.ExecutionLocations executionLocation)
     {
       HttpContext.Current.Items[_clientContextName] = clientContext;
+    }
+
+    private const string _applicationContextName = "Csla.ApplicationContext";
+
+    /// <summary>
+    /// Gets or sets a reference to the current ApplicationContext.
+    /// </summary>
+    public ApplicationContext ApplicationContext
+    {
+      get
+      {
+        return (ApplicationContext)HttpContext.Current.Items[_applicationContextName];
+      }
+      set
+      {
+        HttpContext.Current.Items[_applicationContextName] = value;
+      }
     }
   }
 }
